Login refactor
